+++ Examples/SnowflakeIDGenerator.Example.Web/Program.cs
@@ -1,0 +1,29 @@
+using SnowflakeID;
+
+var builder = WebApplication.CreateBuilder(args);
+
+// Add services to the container.
+
+builder.Services.AddControllers();
+// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+builder.Services.AddSnowflakeIdGeneratorService(builder.Configuration.GetSection(nameof(SnowflakeIdGeneratorOptions)).Get<SnowflakeIdGeneratorOptions>());
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
+
+app.UseHttpsRedirection();
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();