+++ SnowflakeIDGenerator/Snowflake.cs
@@ -1,9 +1,10 @@
-﻿// Copyright (c) 2022-2023, Federico Seckel.
+﻿// Copyright (c) 2022-2024, Federico Seckel.
 // Licensed under the BSD 3-Clause License. See LICENSE file in the project root for full license information.
 
 // Ignore Spelling: Rebase
 
 using SnowflakeID.Exceptions;
+using SnowflakeID.Helpers;
 using System;
 using System.Globalization;
 
@@ -39,14 +40,6 @@ namespace SnowflakeID
         private const ulong MASK_SECUENCIA_RIGHT_ALIGNED
                                                 = 0b0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_0000_1111_1111_1111;
 
-
-#if NETCOREAPP2_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER
-        private static readonly DateTime defaultEpoch = DateTime.UnixEpoch;
-#else
-        // This should be DateTime.UnixEpoch. However, that constant is only available in .netCore and net5 or newer
-        private static readonly DateTime defaultEpoch = new(year: 1970, month: 1, day: 1, hour: 0, minute: 0, second: 0, kind: DateTimeKind.Utc);
-#endif
-
         /// <summary>
         /// Current epoch being used
         /// </summary>
@@ -61,7 +54,7 @@ namespace SnowflakeID
         /// <summary>
         /// Class constructor using default epoch (UNIX time 1-1-1970)
         /// </summary>
-        public Snowflake() : this(defaultEpoch) { }
+        public Snowflake() : this(GlobalConstants.DefaultEpoch) { }
 
         /// <summary>
         /// Class constructor using a custom date as epoch.
@@ -80,6 +73,7 @@ namespace SnowflakeID
         /// <summary>
         /// Gets / Sets machine / server number
         /// </summary>
+        /// <exception cref="ArgumentOutOfRangeException"></exception>
         [CLSCompliant(false)]
         public ulong MachineId
         {
@@ -100,12 +94,17 @@ namespace SnowflakeID
         /// <summary>
         /// Gets / Sets machine / server number
         /// </summary>
+        /// <exception cref="ArgumentOutOfRangeException"></exception>
         public int MachineIdInt32
         {
             get => (int)MachineId;
             set
             {
+#if NET8_0_OR_GREATER
+                ArgumentOutOfRangeException.ThrowIfNegative(value);
+#else
                 if (value < 0) { throw new ArgumentOutOfRangeException(nameof(MachineIdInt32)); }
+#endif
                 MachineId = (ulong)value;
             }
         }
@@ -114,6 +113,7 @@ namespace SnowflakeID
         /// <summary>
         /// Gets / Sets sequence
         /// </summary>
+        /// <exception cref="ArgumentOutOfRangeException"></exception>
         [CLSCompliant(false)]
         public ulong Sequence
         {
@@ -134,12 +134,17 @@ namespace SnowflakeID
         /// <summary>
         /// Gets / Sets machine / server number
         /// </summary>
+        /// <exception cref="ArgumentOutOfRangeException"></exception>
         public int SequenceInt32
         {
             get => (int)Sequence;
             set
             {
+#if NET8_0_OR_GREATER
+                ArgumentOutOfRangeException.ThrowIfNegative(value);
+#else
                 if (value < 0) { throw new ArgumentOutOfRangeException(nameof(SequenceInt32)); }
+#endif
                 Sequence = (ulong)value;
             }
         }
@@ -164,12 +169,17 @@ namespace SnowflakeID
         /// <summary>
         /// Gets / Sets timeStamp as number of milliseconds since selected epoch
         /// </summary>
+        /// <exception cref="ArgumentOutOfRangeException"></exception>
         public long TimestampInt64
         {
             get => (long)Timestamp;
             set
             {
+#if NET8_0_OR_GREATER
+                ArgumentOutOfRangeException.ThrowIfNegative(value);
+#else
                 if (value < 0) { throw new ArgumentOutOfRangeException(nameof(TimestampInt64)); }
+#endif
                 ulong newVal = (ulong)value & MASK_DATETIMEMILLIS_RIGHT_ALIGNED;
                 if (newVal != (ulong)value) { throw new ArgumentOutOfRangeException(nameof(TimestampInt64)); }
                 Timestamp = newVal;
@@ -344,7 +354,7 @@ namespace SnowflakeID
         public int CompareTo(Snowflake other)
         {
             if (other == null) { return 1; }
-            if (Epoch != other.Epoch) { throw new SnowflakesUsingDifferentEpochsException(); }
+            if (Epoch != other.Epoch) { throw new SnowflakesUsingDifferentEpochsException(SnowflakesUsingDifferentEpochsException.DefaultMessage, nameof(other)); }
             return Id.CompareTo(other.Id);
         }
 
@@ -420,7 +430,6 @@ namespace SnowflakeID
         /// </summary>
         /// <param name="s"></param>
         public static explicit operator Snowflake(string s) => Parse(s);
-
 
         /// <summary>
         /// Explicit cast from <see cref="string"/>